+++ src/devcenter/azext_devcenter/tests/latest/helper.py
@@ -51,13 +51,17 @@ def create_dev_center_with_identity(self):
 def create_virtual_network_with_subnet(self):
     self.kwargs.update({
         'vNetName': self.create_random_name(prefix='cli', length=24),
-        'subnetName': self.create_random_name(prefix='cli', length=24)
+        'subnetName': self.create_random_name(prefix='cli', length=24),
+        'nsgName': self.create_random_name(prefix='cli', length=12),
     })
 
     self.cmd(
         'az network vnet create -n "{vNetName}" --location "{location}" -g "{rg}"')
+    
+    self.cmd(
+        'az network nsg create -n "{nsgName}" --location "{location}" -g "{rg}"')
 
-    return self.cmd('az network vnet subnet create -n "{subnetName}" --vnet-name "{vNetName}" -g "{rg}" --address-prefixes "10.0.0.0/21"').get_output_in_json()
+    return self.cmd('az network vnet subnet create --nsg "{nsgName}" -n "{subnetName}" --vnet-name "{vNetName}" -g "{rg}" --address-prefixes "10.0.0.0/21"').get_output_in_json()
 
 
 def create_sig(self):
@@ -70,7 +74,8 @@ def create_sig(self):
         'publisher': "MicrosoftWindowsDesktop",
         'offer': "Windows-10",
         'sku': "win10-21h2-entn-g2",
-        'imageVersion': "1.0.0"
+        'imageVersion': "1.0.0",
+        'nsgName': self.create_random_name(prefix='cli', length=12),
     })
 
     sig = self.cmd(
@@ -84,6 +89,8 @@ def create_sig(self):
         'sigId': sig['id']
     })
 
+    create_virtual_network_with_subnet(self);
+
     # Create compute virtual machine
     self.cmd('az vm create -n "{computeVmName}" '
              '-g "{rg}" '
@@ -91,6 +98,8 @@ def create_sig(self):
              '--location "{location}" '
              '--security-type TrustedLaunch '
              '--admin-password "{computeVmPassword}" '
+             '--vnet-name "{vNetName}" '
+             '--subnet "{subnetName}" '
              '--admin-username "{computeUserName}"')
 
     compute_vm = self.cmd('az vm show -n "{computeVmName}" '
@@ -164,6 +173,39 @@ def create_network_connection(self):
         'networkConnectionId': network_connection['id'],
     })
 
+def create_network_connection_dp(self):
+    subnet = create_virtual_network_with_subnet_dp(self)
+
+    self.kwargs.update({
+        'subnetId': subnet['id'],
+        'networkConnectionName': self.create_random_name(prefix='cli', length=24),
+        'networkingRgName1': self.create_random_name(prefix='cli', length=24),
+    })
+
+    network_connection = self.cmd('az devcenter admin network-connection create '
+                                  '--location "{location}" '
+                                  '--tags CostCode="12345" '
+                                  '--name "{networkConnectionName}" '
+                                  '--domain-join-type "AzureADJoin" '
+                                  '--subnet-id "{subnetId}" '
+                                  '--networking-resource-group-name "{networkingRgName1}" '
+                                  '--resource-group "{rg}"').get_output_in_json()
+
+    self.kwargs.update({
+        'networkConnectionId': network_connection['id'],
+    })
+
+def create_virtual_network_with_subnet_dp(self):
+    self.kwargs.update({
+        'vNetName': self.create_random_name(prefix='cli', length=24),
+        'subnetName': self.create_random_name(prefix='cli', length=24)
+    })
+
+    self.cmd(
+        'az network vnet create -n "{vNetName}" --location "{location}" -g "{rg}"')
+
+    return self.cmd('az network vnet subnet create -n "{subnetName}" --vnet-name "{vNetName}" -g "{rg}" --address-prefixes "10.0.0.0/21"').get_output_in_json()
+
 
 def create_attached_network_dev_box_definition(self):
     self.kwargs.update({
@@ -174,7 +216,7 @@ def create_attached_network_dev_box_definition(self):
     create_project(self)
     create_network_connection(self)
     self.kwargs.update({
-        'imageRefId': "/subscriptions/{subscriptionId}/resourceGroups/{rg}/providers/Microsoft.DevCenter/devcenters/{devcenterName}/galleries/default/images/microsoftwindowsdesktop_windows-ent-cpc_win11-21h2-ent-cpc-m365",
+        'imageRefId': "/subscriptions/{subscriptionId}/resourceGroups/{rg}/providers/Microsoft.DevCenter/devcenters/{devcenterName}/galleries/default/images/microsoftwindowsdesktop_windows-ent-cpc_win11-22h2-ent-cpc-os",
         'devBoxDefinitionName': self.create_random_name(prefix='c1', length=12),
         'osStorageType': "ssd_1024gb",
         'skuName': "general_a_8c32gb_v1",
@@ -292,7 +334,7 @@ def create_pool(self):
 
 def create_pool_dataplane_dependencies(self):
     
-    create_network_connection(self)
+    create_network_connection_dp(self)
 
     self.kwargs.update({
         'imageRefId': "/subscriptions/{subscriptionId}/resourceGroups/{rg}/providers/Microsoft.DevCenter/devcenters/{devcenterName}/galleries/default/images/microsoftwindowsdesktop_windows-ent-cpc_win11-21h2-ent-cpc-m365",
@@ -431,7 +473,7 @@ def create_dev_box_dependencies(self):
     create_dev_center(self)
     create_project(self)
     add_dev_box_user_role_to_project(self)
-    create_network_connection(self)
+    create_network_connection_dp(self)
 
     self.kwargs.update({
         'imageRefId': "/subscriptions/{subscriptionId}/resourceGroups/{rg}/providers/Microsoft.DevCenter/devcenters/{devcenterName}/galleries/default/images/microsoftwindowsdesktop_windows-ent-cpc_win11-22h2-ent-cpc-os",