+++ src/quantum/azext_quantum/tests/latest/test_quantum_workspace.py
@@ -108,11 +108,34 @@ class QuantumWorkspacesScenarioTest(ScenarioTest):
             self.check("provisioningState", "Deleting")
             ])
 
-            # Repeat the tests without the "--skip-role-assignment" parameter
+            # Create workspace with "--skip-role-assignment" and "--skip-autoadd" parameters
             test_workspace_temp = get_test_workspace_random_name()
+            self.cmd(f'az quantum workspace create --skip-autoadd -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account} -r {test_provider_sku_list} -o json --skip-role-assignment', checks=[
+            self.check("name", test_workspace_temp),
+            self.check("provisioningState", "Accepted")  # Status is accepted since we're not linking the storage account.
+            ])
 
-            # create
+            # delete
+            self.cmd(f'az quantum workspace delete -g {test_resource_group} -w {test_workspace_temp} -o json', checks=[
+            self.check("name", test_workspace_temp),
+            self.check("provisioningState", "Deleting")
+            ])
+
+            # Repeat without the "--skip-role-assignment" or "--skip-autoadd" parameters (Uses ARM template and adds C4A plans)
+            test_workspace_temp = get_test_workspace_random_name()
             self.cmd(f'az quantum workspace create -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account} -r {test_provider_sku_list} -o json', checks=[
+            self.check("name", DEPLOYMENT_NAME_PREFIX + test_workspace_temp),
+            ])
+
+            # delete
+            self.cmd(f'az quantum workspace delete -g {test_resource_group} -w {test_workspace_temp} -o json', checks=[
+            self.check("name", test_workspace_temp),
+            self.check("provisioningState", "Deleting")
+            ])
+
+            # Create a workspace specifying "--skip-autoadd"
+            test_workspace_temp = get_test_workspace_random_name()
+            self.cmd(f'az quantum workspace create --skip-autoadd -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account} -r {test_provider_sku_list} -o json', checks=[
             self.check("name", DEPLOYMENT_NAME_PREFIX + test_workspace_temp),
             ])
 
@@ -124,9 +147,7 @@ class QuantumWorkspacesScenarioTest(ScenarioTest):
 
             # Create a workspace specifying a storage account that is not Standard_LRS
             test_workspace_temp = get_test_workspace_random_name()
-
-            # create
-            self.cmd(f'az quantum workspace create -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account_grs} -r {test_provider_sku_list} -o json', checks=[
+            self.cmd(f'az quantum workspace create --skip-autoadd -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account_grs} -r {test_provider_sku_list} -o json', checks=[
             self.check("name", DEPLOYMENT_NAME_PREFIX + test_workspace_temp),
             ])
 
@@ -138,9 +159,7 @@ class QuantumWorkspacesScenarioTest(ScenarioTest):
 
             # Create a workspace with a maximum length name, but make sure the deployment name was truncated to a valid length
             test_workspace_temp = get_test_workspace_random_long_name()
-
-            # create
-            self.cmd(f'az quantum workspace create -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account_grs} -r {test_provider_sku_list} -o json', checks=[
+            self.cmd(f'az quantum workspace create --skip-autoadd -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account_grs} -r {test_provider_sku_list} -o json', checks=[
             self.check("name", (DEPLOYMENT_NAME_PREFIX + test_workspace_temp)[:64]),
             ])
 
@@ -230,12 +249,3 @@ class QuantumWorkspacesScenarioTest(ScenarioTest):
         workspace_location = None
         _autoadd_providers(cmd, providers_in_region, providers_selected, workspace_location, True)
         assert providers_selected[0] == {"provider_id": "foo", "sku": "foo_credits_for_all_plan", "offer_id": "foo_offer", "publisher_id": "foo0123456789"}
-
-        # Make sure we get an error message if there are no auto_add providers and providers_selected is empty, like when there's no -r in the command:
-        try:
-            test_provider.properties.skus[0].auto_add = False
-            providers_selected = []
-            _autoadd_providers(cmd, providers_in_region, providers_selected, workspace_location, True)
-            assert False
-        except RequiredArgumentMissingError as e:
-            assert str(e) == "A list of Azure Quantum providers and SKUs is required."